renamed media-selection container to multi-media-selection
