+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/Form/fields/MediaSelection.js
@@ -2,8 +2,8 @@
 import React from 'react';
 import {observer} from 'mobx-react';
 import type {FieldTypeProps} from 'sulu-admin-bundle/types';
+import MultiMediaSelectionComponent from '../../MultiMediaSelection';
 import type {Value} from '../../MediaSelection/types';
-import MediaSelectionComponent from '../../MediaSelection';
 
 @observer
 export default class MediaSelection extends React.Component<FieldTypeProps<Value>> {
@@ -24,7 +24,7 @@ export default class MediaSelection extends React.Component<FieldTypeProps<Value
         const {locale} = formInspector;
 
         return (
-            <MediaSelectionComponent
+            <MultiMediaSelectionComponent
                 disabled={!!disabled}
                 locale={locale}
                 onChange={this.handleChange}
