--- src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelection/README.md
@@ -1,1 +1,0 @@
-The `MediaSelection` is a content-type based on the `MultiItemSelection` component. By using the `MediaSelection` you can assign media entries inside a form.
