--- src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelection/index.js
@@ -1,4 +1,0 @@
-// @flow
-import MediaSelection from './MediaSelection';
-
-export default MediaSelection;
