+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MultiMediaSelection/README.md
@@ -1,0 +1,1 @@
+The `MultiMediaSelection` is a content-type based on the `MultiItemSelection` component. By using the `MediaSelection` you can assign media entries inside a form.
