+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MultiMediaSelection/index.js
@@ -1,0 +1,4 @@
+// @flow
+import MultiMediaSelection from './MultiMediaSelection';
+
+export default MultiMediaSelection;
